# docs/API.yaml
openapi: 3.1.0
info:
  title: Wyvern Drive API
  description: |
    Internal API documentation for Wyvern Drive Supabase Edge Function.

    **⚠️ INTERNAL USE ONLY - NOT FOR PUBLIC DISTRIBUTION**

    This API handles file metadata storage, versioning, sharing, and video streaming.
    Actual file data is stored on Discord CDN; this API manages the index.
  version: 1.0.0
  contact:
    name: Developer
servers:
  - url: https://{project-ref}.supabase.co/functions/v1/api
    description: Production (Supabase Edge Function)
    variables:
      project-ref:
        default: your-project-ref
...
paths:
...
  /refresh-urls:
    post:
      tags: [Utilities]
      summary: Batch refresh expired Discord URLs
      description: |
        Refreshes expired Discord CDN URLs for file chunks.
        Uses webhook-based refresh (primary) or bot token fallback.
        Only chunks with message ID (m) and channel ID (cid) can be refreshed.
      operationId: refreshUrls
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chunks]
              properties:
                chunks:
                  type: array
                  description: Chunks needing URL refresh
                  items:
                    type: object
                    required: [i, u]
                    properties:
                      i:
                        type: integer
                        description: Chunk index
                      u:
                        type: string
                        description: Current (possibly expired) Discord CDN URL
                      m:
                        type: string
                        description: Discord message ID (required for refresh)
                      cid:
                        type: string
                        description: Discord channel ID (required for refresh)
                webhookUrl:
                  type: string
                  description: Optional webhook URL for refresh (uses user's webhook)
      responses:
        '200':
          description: Refreshed URLs
          content:
            application/json:
              schema:
                type: object
                properties:
                  refreshed:
                    type: object
                    additionalProperties:
                      type: string
                    description: Map of chunk index to new URL
                    example:
                      "0": "https://cdn.discordapp.com/attachments/..."
                      "1": "https://cdn.discordapp.com/attachments/..."
        '400':
          description: Invalid chunks array
        '500':
          $ref: '#/components/responses/ServerError'
...
  /shares/{userId}/{fileId}:
    get:
      tags: [Shares]
      summary: List shares for file
      operationId: listShares
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/fileId'
      responses:
        '200':
          description: List of share links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Share'

    post:
      tags: [Shares]
      summary: Create share link
      description: Cannot share encrypted files.
      operationId: createShare
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/fileId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                expiresIn:
                  type: integer
                  description: Hours until expiry
                password:
                  type: string
                  description: Optional access password
      responses:
        '200':
          description: Created share
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  url:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
                    nullable: true
  /shares/{userId}/{shareId}:
    delete:
      tags: [Shares]
      summary: Revoke share link
      operationId: deleteShare
      parameters:
        - $ref: '#/components/parameters/userId'
        - name: shareId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          $ref: '#/components/responses/Success'
  /share/{shareId}/info:
    get:
      tags: [Shares]
      summary: Get share info (public)
      description: No authentication required. Returns metadata only.
      operationId: getShareInfo
      parameters:
        - name: shareId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Share metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  fileName:
                    type: string
                  fileSize:
                    type: integer
                  fileType:
                    type: string
                  expiresAt:
                    type: string
                    nullable: true
                  passwordRequired:
                    type: boolean
                  downloadCount:
                    type: integer
        '404':
          description: Share not found
        '410':
          description: Share expired
  /share/{shareId}:
    get:
      tags: [Shares]
      summary: Download shared file (public)
      description: No authentication. Password via query param if required.
      operationId: downloadShare
      parameters:
        - name: shareId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: password
          in: query
          schema:
            type: string
      responses:
        '200':
          description: File download
          content:
            application/octet-stream: {}
        '401':
          description: Password required
        '410':
          description: Share expired
components:
  parameters:
    userId:
      name: userId
      in: path
      required: true
      description: User identifier (hashed webhook URL)
      schema:
        type: string
    fileId:
      name: fileId
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Share:
      type: object
      properties:
        id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          nullable: true
        download_count:
          type: integer
  responses:
    Success:
      description: Operation successful
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string